use nint instead of int when rounding obs_ii and jj this prevents some mislocation in xb_to_radiance

+++ EnKF/src/xb.f
@@ -1078,8 +1078,8 @@ subroutine xb_to_radiance(inputfile,proj,ix,jx,kx,xlong,xlat,landmask,iob_radmin
   do iob = ystart, yend
      obs_ii=lon_radiance(iob)
      obs_jj=lat_radiance(iob)
-     x = int( obs_ii )
-     y = int( obs_jj )
+     x = nint( obs_ii )
+     y = nint( obs_jj )
 
   ! 4a3. Converting WRF data for CRTM structure
   ! --------------------------------
@@ -1297,8 +1297,8 @@ subroutine xb_to_radiance(inputfile,proj,ix,jx,kx,xlong,xlat,landmask,iob_radmin
   do iob = iob_radmin, iob_radmax
       obs_ii=obs%position(iob,1)
       obs_jj=obs%position(iob,2)
-      i1=int(obs_ii)
-      j1=int(obs_jj)
+      i1=nint(obs_ii)
+      j1=nint(obs_jj)
       dx  = obs_ii-real(i1)
       dxm = real(i1+1)-obs_ii
       dy  = obs_jj-real(j1)
@@ -1319,8 +1319,8 @@ subroutine xb_to_radiance(inputfile,proj,ix,jx,kx,xlong,xlat,landmask,iob_radmin
     do iob = iob_radmin, iob_radmax
       obs_ii=obs%position(iob,1)
       obs_jj=obs%position(iob,2)
-      x = int( obs_ii )
-      y = int( obs_jj )
+      x = nint( obs_ii )
+      y = nint( obs_jj )
       if (Sensor_Id == 'abi_gr' ) then
          if (obs%ch(iob) .eq. 8) xb_tb(iob) = Tb(x,y,1) !6.19um
          if (obs%ch(iob) .eq. 9) xb_tb(iob) = Tb(x,y,2) !6.95um
